backend/requirements.txt
flask==3.1.*
flask-cors==5.*
readability-lxml==0.8.*
beautifulsoup4==4.*
requests==2.*
pypandoc==1.*
pathvalidate==3.*
google-api-python-client==2.*
google-auth-oauthlib==1.*
lxml[html_clean]==5.*
Pillow==11.*
mathml-to-latex==1.*
